Simplification to speed up pre-processing of microstructures
The generate_microstructures.py script has been updated to more efficiently calculate "kosher" grain centroids. DREAM.3D can generate microstructures as fully periodic in all direction or not periodic at all. In the former case, grains that are "split" by the microstructure boundary and appear on both sides (i.e., faces) of the microstructure have a reported and calculated grain centroid that is not realistic, i.e., in the center of the microstructure. Before the generate_microstructures.py script can band grains, a realistic grain centroid must be determined. This is done in the "kosher_centroids" function. Elements in grains are shifted to account for grains split by the microstructure boundary and these hypothetical element centroids are then used to calculate grain centroids that are more realistic.

This function previously manipulated the grain centroids of ALL grains in a microstructure, and this did not previously result in a bottleneck because of the relatively small microstructures analyzed (e.g., 30 x 30 x 30 elements). When much larger microstructures are generated, the process is substantially slowed. The "kosher_centroids" function has been simplified to only consider this task for "surface feature" grains, i.e., grains that have at least one element at the surface of the microstructure, as generated by DREAM.3D.

For a microstructure with 30 x 30 x 30 elements and ~280 grains that is fully periodic, this change reduced the time to calculate the "kosher" grain centroids from 1.27 to 0.98 seconds. For a much larger microstructure with 90 x 90 x 90 and ~7,500 grains, this change reduced this time from 22.3 to 10.3 minutes. Substantially larger microstructures will further benefit from this code change..

The reason that grain centroids are manipulated in this manner is to appropriately consider the periodic microstructure as generated by DREAM.3D and to appropriately analyze grains near the microstructure boundary. Of course, users can substantially reduce the cost of this pre-processing steps if desired (by setting "compute_kosher_grain_centroids" to False), but this may have substantial effects on the computed extreme value fatigue response. The files necessary to execute PRISMS-Plasticity simulations are generated directly from the DREAM.3D portion of the code, so users can return to these considerations after the CPFEM simulations finish. This is also true for the sub banding process, which takes some time to complete and that can be disabled by setting "create_sub_bands" to False (FIPs can still be averaged over grains or bands, or users can create the sub bands at a later time). Please contact Kris at stopka.kris@gmail.com for more details.