Scripts for 3rd PRISMS-Fatigue manuscript
These scripts are associated with the third PRISMS-Fatigue manuscript entitled "Microstructure effects on the extreme value fatigue response of FCC metals and alloys: Effects of sample size and grain neighborhood".

Also, small change to generate_microstructures.py so that users can specify the volume percent of sub-bands as a percent of total grain volume. The default is set to 10%. This is especially important when comparing FIPs from microstructures with different grain sizes!