Fixed files
Updated .json that sometimes caused pipeline to crash. The fix includes separate .dream3d files located in the "generate_simple_microstructure" folder.

This was because the .dream3d files located in the src folder have two phases specified, whereas this script works for a single phase.

This may be updated in the future.