Update links to papers, slight update to calculate_FIPs.py
Updated links to associated manuscripts under applications folder.

In the calculate_FIPs.py script, the code now reads the results .csv files and determines whether there is a header column. If so, this column is overwritten. This will make it easy to read in the .csv files from the different PRISMS-Fatigue data sets available on Materials Commons that are part of the manuscripts in the applications folder. Users need only to keep track of what values are in each column when defining a new FIP or using existing values in the .csv files.